Get uniswap events in block range (#139)
* Get events in block range.

* Get uniswap events in block range.

+++ packages/address-watcher/src/entity/BlockProgress.ts
@@ -6,13 +6,13 @@ export class BlockProgress {
   @PrimaryColumn('varchar', { length: 66 })
   blockHash!: string;
 
-  @Column('numeric')
+  @Column('integer')
   blockNumber!: number;
 
-  @Column('numeric')
+  @Column('integer')
   numTx!: number;
 
-  @Column('numeric')
+  @Column('integer')
   numTracedTx!: number;
 
   @Column('boolean')
